refactor(search): split OR logic into orOptimizations and orQueryHandler modules
